Better Image Handling & FadeIn Effects, Placeholders Moved To Amazon, Tests With Update & Delete Functions
We fixed the way error images are handled and allowing javascript to detect all the non-loaded images and replacing their src with a link to a placeholder showing that and image should be present. The image styles were updated in theeir blade files to allow for animation transitions that smoothly FadeIn the all pictures. Most placeholders for (video/images) were moved from local storage to Amazon and the files were updated. We tried to allow for image uploads via TinyMCE to Amazon, however the job is not yet complete. We added new scriptsto allow for tests to be updated as well as deleted.