Admin can now enroll client from their profiles, fixes for lost images
The admin now has the ability to not only profile clients but can now enroll them directly from their profile page. We also worked on fixing the ability of javascript to detect lost or not found images and replace them with placeholders instead of just the default broken image symbol.